Split the serializer and parser -> Both was created by Renette I just split them.
